separated extensions, fixed all marks
