--- Lynked/Currency+Extension.swift
@@ -1,42 +1,0 @@
-//
-//  Currency+Extension.swift
-//  Lynked
-//
-//  Created by Matthew Singleton on 1/24/17.
-//  Copyright © 2017 Matthew Singleton. All rights reserved.
-//
-
-import Foundation
-import UIKit
-
-
-extension String {
-  
-  // formatting text for currency textField
-  func currencyInputFormatting() -> String {
-    
-    var number: NSNumber!
-    let formatter = NumberFormatter()
-    formatter.numberStyle = .currencyAccounting
-    formatter.currencySymbol = "$"
-    formatter.maximumFractionDigits = 2
-    formatter.minimumFractionDigits = 2
-    
-    var amountWithPrefix = self
-    
-    // remove from String: "$", ".", ","
-    let regex = try! NSRegularExpression(pattern: "[^0-9]", options: .caseInsensitive)
-    amountWithPrefix = regex.stringByReplacingMatches(in: amountWithPrefix, options: NSRegularExpression.MatchingOptions(rawValue: 0), range: NSMakeRange(0, self.characters.count), withTemplate: "")
-    
-    let double = (amountWithPrefix as NSString).doubleValue
-    number = NSNumber(value: (double / 100))
-    
-    // if first number is 0 or all numbers were deleted
-    guard number != 0 as NSNumber else {
-      return ""
-    }
-    
-    return formatter.string(from: number)!
-  }
-  
-}
