+++ Lynked/NSObject+Extensions.swift
@@ -1,0 +1,17 @@
+//
+//  NSObject+Extensions.swift
+//  Lynked
+//
+//  Created by Matthew Howes Singleton on 6/18/17.
+//  Copyright © 2017 Matthew Singleton. All rights reserved.
+//
+
+import Foundation
+import UIKit
+
+extension NSObject {
+    func delay(_ delay:Double, closure:@escaping ()->()) {
+        let when = DispatchTime.now() + delay
+        DispatchQueue.main.asyncAfter(deadline: when, execute: closure)
+    }
+}
