--- Lynked/ShadowForTableViewCell.swift
@@ -1,37 +1,0 @@
-//
-//  ShadowForTableViewCell.swift
-//  Lynked
-//
-//  Created by Matthew Singleton on 12/14/16.
-//  Copyright © 2016 Matthew Singleton. All rights reserved.
-//
-
-import UIKit
-
-private var materialKey = false
-
-
-extension UIView {
-  
-  @IBInspectable var cellDesign: Bool {
-    
-    get {
-      
-      return materialKey
-      
-    } set {
-      
-      materialKey = newValue
-      
-      if materialKey {
-        
-        self.layer.masksToBounds = false
-        self.layer.cornerRadius = 3.0
-        self.layer.shadowOpacity = 0.8
-        self.layer.shadowRadius = 3.0
-        self.layer.shadowOffset = CGSize(width: 0.0, height: 2.0)
-        self.layer.shadowColor = UIColor(red: 157.0/255.0, green: 157.0/255.0, blue: 157.0/255.0, alpha: 1.0).cgColor
-      }
-    }
-  }
-}
