+++ Lynked/UIViewController+Extensions.swift
@@ -1,0 +1,36 @@
+//
+//  UIViewController+Extensions.swift
+//  Lynked
+//
+//  Created by Matthew Howes Singleton on 6/18/17.
+//  Copyright © 2017 Matthew Singleton. All rights reserved.
+//
+
+import Foundation
+import UIKit
+
+extension UIViewController {
+    func showAlertWith(title: String?, message: String?) {
+        let alert = UIAlertController(title: title, message: message, preferredStyle: .alert)
+        let cancelAction = UIAlertAction(title: "Dismiss", style: .cancel, handler: nil)
+        alert.addAction(cancelAction)
+        present(alert, animated: true, completion: nil)
+    }
+    
+    func setNavBar() {
+        self.navigationController?.isNavigationBarHidden = false
+        navigationController?.navigationBar.barTintColor = UIColor(red: 108.0/255.0,
+                                                                   green: 158.0/255.0,
+                                                                   blue: 236.0/255.0,
+                                                                   alpha: 1.0)
+        UINavigationBar.appearance().tintColor = .white
+        UINavigationBar.appearance().titleTextAttributes = [NSForegroundColorAttributeName : UIColor.white]
+        navigationController?.navigationBar.titleTextAttributes = [NSForegroundColorAttributeName: UIColor.white,
+                                                                   NSFontAttributeName: UIFont(name: "GillSans-Bold",
+                                                                                               size: 18)!]
+    }
+    
+    func dismissKeyboard() {
+        view.endEditing(true)
+    }
+}
